usr.bin/cc: a prototype cleaning up
